refactor: 숫자 뽑기 함수 파일 분리, 게임 종료 함수명 변경
